Create separate subreddit model
Created subreddit model
Removed description column
Used promise.all to fetch posts
Edited audienceQueries to include related subreddits
Removed popular and new subreddit router and controller

+++ prisma/queries/audienceQueries.js
@@ -1,12 +1,26 @@
 const { prisma } = require("../prismaClient");
 
-const createAud = async (title, subreddits, userId, searchTerm) => {
+const createAud = async (title, subreddits, userId, searchTerm = "") => {
   const audience = await prisma.audience.create({
     data: {
       title,
-      subreddits,
       userId,
       searchTerm,
+      subreddits: {
+        connectOrCreate: subreddits.map((sub) => ({
+          where: { name: sub.name },
+          create: {
+            name: sub.name,
+            title: sub.title,
+            subscribers: sub.subscribers,
+            icon: sub.icon,
+            url: sub.url,
+          },
+        })),
+      },
+    },
+    include: {
+      subreddits: true,
     },
   });
 
@@ -15,29 +29,73 @@ const createAud = async (title, subreddits, userId, searchTerm) => {
 };
 
 const getAllAudiences = async (userId) => {
-  const audiences = await prisma.audience.findMany({ where: { userId } });
+  const audiences = await prisma.audience.findMany({
+    where: { userId },
+    include: { subreddits: true },
+  });
   console.log(audiences);
   return audiences;
 };
 
 const getOneAudience = async (id) => {
-  const audience = await prisma.audience.findUnique({ where: { id } });
+  const audience = await prisma.audience.findUnique({
+    where: { id },
+    include: { subreddits: true },
+  });
   console.log("single audience", audience);
   return audience;
 };
 
 const delAud = async (id) => {
+  const deletedAud = await prisma.audience.findUnique({
+    where: { id },
+    include: { subreddits: true },
+  });
+
+  if (!deletedAud) {
+    return deletedAud;
+  }
+
+  const subredditIds = deletedAud.subreddits.map((sub) => ({ id: sub.id }));
+  await prisma.audience.update({
+    where: { id },
+    data: {
+      subreddits: {
+        disconnect: subredditIds,
+      },
+    },
+  });
   const audience = await prisma.audience.delete({
     where: { id },
+    include: { subreddits: true },
   });
-  console.log("deleted audience", audience);
+
+  console.log("Deleted audience and related subreddits", audience);
   return audience;
 };
 
 const updateAud = async (id, title, subreddits) => {
   const audience = await prisma.audience.update({
     where: { id },
-    data: { title, subreddits },
+    data: {
+      title,
+      subreddits: {
+        connectOrCreate: subreddits.map((sub) => ({
+          where: { name: sub.name },
+          create: {
+            name: sub.name,
+            title: sub.title,
+            subscribers: Number(sub.subscribers),
+            icon: sub.icon,
+            url: sub.url,
+          },
+        })),
+        set: subreddits.map((sub) => ({ name: sub.name })),
+      },
+    },
+    include: {
+      subreddits: true,
+    },
   });
 
   console.log("updated audience", audience);
